+++ 3D_Base/Source/GameObject/MeshObject/StaticMesh/PlayerBase/PlayerState/PlayerActionState/PlayerHandWhiffState/CPlayerHandWhiffState.cpp
@@ -44,7 +44,7 @@ void CPlayerHandWhiffState::Enter()
 	m_StartQuat = m_pPlayer.GetQuaternion();
 
 	//攻撃の開始時間を取得.
-	m_StartTime = static_cast<float>(CTimeManager::GetTotalTime());
+	m_StartTime = CTimeManager::GetTotalTime();
 
 	//手の開始位置を設定.
 	m_RightHandStartPos = m_pPlayer.GetPlayerRightHand().GetPosition();
