+++ 3D_Base/Source/GameObject/MeshObject/StaticMesh/PlayerBase/PlayerState/PlayerActionState/PlayerKnockbackState/CPlayerKnockbackState.cpp
@@ -37,7 +37,7 @@ void CPlayerKnockbackState::Enter()
 	AssetManager::Sound()->PlaySE(enSoundList::SE_Knockback);
 	
 	//攻撃の開始時間を取得.
-	m_StartTime = static_cast<float>(CTimeManager::GetTotalTime());
+	m_StartTime = CTimeManager::GetTotalTime();
 
 	//プレイヤーの位置を取得.
 	D3DXVECTOR3 playerPos = m_pPlayer.GetPosition();
